Move config api creds to opencast
