+++ backend/src/config/opencast.rs
@@ -1,6 +1,8 @@
 use std::{str::FromStr, fmt};
 
 use hyper::Uri;
+use base64::Engine as _;
+use secrecy::{ExposeSecret as _, Secret};
 use serde::Deserialize;
 
 use crate::{
@@ -49,6 +51,15 @@ pub(crate) struct OpencastConfig {
     /// be trusted.
     #[config(default = [])]
     pub(crate) other_hosts: Vec<HttpHost>,
+
+    /// Username of the user used to communicate with Opencast for data syncing
+    /// and external API authentication.
+    /// This user has to have access to all events and series. Currently, that
+    /// user has to be admin.
+    pub user: String,
+
+    /// Password of the user used to communicate with Opencast.
+    password: Secret<String>,
 }
 
 impl OpencastConfig {
@@ -109,6 +120,13 @@ impl OpencastConfig {
                 .unwrap();
             ToolBaseUri(uri)
         })
+    }
+
+    pub(crate) fn basic_auth_header(&self) -> Secret<String> {
+        let credentials = format!("{}:{}", self.user, self.password.expose_secret());
+        let encoded_credentials = base64::engine::general_purpose::STANDARD.encode(credentials);
+        let auth_header = format!("Basic {}", encoded_credentials);
+        Secret::new(auth_header)
     }
 
     fn unwrap_host(&self) -> &HttpHost {
