+++ backend/src/sync/mod.rs
@@ -1,5 +1,3 @@
-use base64::Engine as _;
-use secrecy::{ExposeSecret as _, Secret};
 use core::fmt;
 use std::time::Duration;
 
@@ -41,14 +39,6 @@ pub(crate) async fn check_compatibility(client: &OcClient) -> Result<()> {
 
 #[derive(Debug, confique::Config)]
 pub(crate) struct SyncConfig {
-    /// Username of the user used to communicate with Opencast for data syncing.
-    /// This user has to have access to all events and series. Currently, that
-    /// user has to be admin.
-    user: String,
-
-    /// Password of the user used to communicate with Opencast.
-    password: Secret<String>,
-
     /// A rough estimate of how many items (events & series) are transferred in
     /// each HTTP request while harvesting (syncing) with the Opencast
     /// instance.
@@ -82,15 +72,6 @@ pub(crate) struct SyncConfig {
     /// load on Opencast, increase to speed up download a bit.
     #[config(default = 8)]
     concurrent_download_tasks: u8,
-}
-
-impl SyncConfig {
-    pub(crate) fn basic_auth_header(&self) -> Secret<String> {
-        let credentials = format!("{}:{}", self.user, self.password.expose_secret());
-        let encoded_credentials = base64::engine::general_purpose::STANDARD.encode(credentials);
-        let auth_header = format!("Basic {}", encoded_credentials);
-        Secret::new(auth_header)
-    }
 }
 
 /// Version of the Tobira-module API in Opencast.
